Changed ListStore to its own class and cleaned Gtk DetailedList.
